Multiple changes made
added in location buttons into the UI
Player can now travel between 3 locations
- Office (Glue Investigations)
- Drive in (Bearly Pictures)
- Finger Monster Alley (Shelf Street)

Removed the purple cubes for travelling within a location. Larger colliders to allow players to require less accuracy when moving between zones
Projector room redone for the minute. Added in the actual projector model and added in a sticker version of Grizzly's stuffing onto the back.

Added controls for camera's on a track and dolly. Currently moving cameras tracking offset to allow the player to look up and down when on a Left/Right Track. Will do the opposite for Forward/Back Tracks.