Suspect Page Finished
Finished the Suspects page for the UI.
When talking to an NPC for the first time a new entry is added in the suspect pool
the new entry is a button. Clicking it opens the suspect data page, which shows the character mugshot and any notes that the player has on the character.